Fixed naming conventions to comply with standard
